Submit D.I.Y. Punks V1 Collection
D.I.Y Punks is the fist collection created using the Generative Art ORC-721 protocol. The collection pieces are inscribed using json files referencing an inscribed spritesheet as the image source. Similar to Ordibots, we are hopeful you will be able to use to protocol to display our images, or we are happy to provide a composite or individual image files for each item in the collection. Note - all of these pieces were created by the inscriber at time of inscribe - more collections have already been completed and minted out - and we are optimistic that OW will be the first to host Gen Art ORC-721 collections! Please follow up with additional questions.